Move load_fm out of the set-difference oracle into spliceai/load_model
load_fm is model loading, orthogonal to the generic SetDifferenceOracle combinator, so
it belongs with load_spliceai/load_lssi. set_difference.py is now pure (no torch/os);
FM_TRACED_DIR + load_fm live in spliceai/load_model.py, and TestLoadFM moves to
tests/test_load_fm.py. No behavior change.
